n-obj-cleanup
For get objects, try with mono color first, and fall back to multi color.
The idea is that whatever invariant the rule is trying to establish will be harder with mono-color than multi-color.
If instead one starts with multi-color where it's intended to be mono color, than the examples might go through by chance, and the test fail: it happens in https://arcprize.org/play?task=1f85a75f where in both modes, the large object ends up being the same.